# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger QR code - OpenAPI 3.0
  version: 1.0.0
servers:
  - url: https://qrc-t8zw.onrender.com
tags:
  - name: qrcode
    description: Access to QR code
  - name: user
    description: Operations about user
paths:
  /qr/user:
    get:
      tags:
        - qrcode
      summary: Returns QR codes by user
      security:
        - authorization: []
      responses:
        '200':
          description: A list of QR codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QRcode'
        '401':
          description: Access token is missing or invalid
        '500':
          description: Internal server error
    post:
      tags:
        - qrcode
      summary: Place a new QR code
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QRcode'
      security:
        - authorization: []
      responses:
        '200':
          description: successful operation
        '401':
          description: Access token is missing or invalid
        '500':
          description: Internal server error
  /qr/g/{id}:
    get:
      tags:
        - qrcode
      summary: Redirect to URL by QR code ID
      parameters:
        - name: id
          in: path
          description: QR code ID
          required: true
          schema:
            type: string
            example: 454c135b-d2c1-4612-8b7f-6aa3e1605ba6
      responses:
        '301':
          description: Redirect to URL by QR code ID
        '500':
          description: Internal server error
  /qr/stat/{id}:
    get:
      tags:
        - qrcode
      summary: Get statistics by QR code ID
      parameters:
        - name: id
          in: path
          description: QR code ID
          required: true
          schema:
            type: string
            example: 454c135b-d2c1-4612-8b7f-6aa3e1605ba6
      responses:
        '200':
          description: A list of QR codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Statistics'
        '500':
          description: Internal server error
  /user/signin:
    post:
      tags:
        - user
      summary: User sign in
      requestBody:
        description: User sign in
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authorized
        '500':
          description: Internal server error
components:
  schemas:
    User:
      type: object
      required:
        - email
        - pass
      properties:
        email:
          type: string
          example: name@mail.com
        pass:
          type: string
          example: 'skjrhwery9.ewerbwcx$'
    QRcode:
      required:
        - from
        - to
      type: object
      properties:
        user_id:
          type: string
          example: 454c135b-d2c1-4612-8b7f-6aa3e1605ba6
        from:
          type: string
          example: 454c135b-d2c1-4612-8b7f-6aa3e1605ba6
        to:
          type: string
          example: https://getbootstrap.com/docs/5.3/examples
        createdAt:
          type: string
          format: date-time
    Statistics:
      type: object
      properties:
        id:
          type: string
          example: 454c135b-d2c1-4612-8b7f-6aa3e1605ba6
        from:
          type: string
          example: 454c135b-d2c1-4612-8b7f-6aa3e1605ba6
        createdAt:
          type: string
          format: date-time
  securitySchemes:
    authorization:
      type: http
      scheme: bearer
      bearerFormat: JWT